Remove the old one
