+++ org/README.md
@@ -1,6 +1,6 @@
 # Istio Organization
 
-The [`istio.yaml`](istio.yaml) file contains configuration for the Istio Github Organization. Data here is automatically synced with Github using [peribolos](https://github.com/kubernetes/test-infra/tree/master/prow/cmd/peribolos).
+The folder contains configuration for the Istio Github Organization. Data here is automatically synced with Github using [peribolos](https://github.com/kubernetes/test-infra/tree/master/prow/cmd/peribolos).
 
 ## Making Changes
 
@@ -8,4 +8,8 @@ To modify the org, simply change the config file and submit a PR. Once the PR is
 
 Changes can be tested with `make test`.
 
-To add yourself to the Org, you just need to add your name to the `members` list.
+To add yourself to the Org, please add your name under `members.yaml` and fill out the form in the PR description.
+
+To add yourself as a developer, which is given additional access to repos, add yourself under `developers.yaml` instead.
+
+Teams, such as working groups and release managers, are controlled by the `teams.yaml` file.
