+++ org/members.yaml
@@ -52,7 +52,6 @@ members:
 - shwsingh
 - ssuchter
 - sunshine6
-- thangbn
 - truongnh1992
 - voanhduy1512
 - wusphinx
