+++ prow/sync-org.sh
@@ -17,10 +17,23 @@
 WD=$(dirname "$0")
 WD=$(cd "$WD"; pwd)
 
-set -e
+set -eu
+
+OUT_DIR=$(mktemp -d)
 
 # Sync the Github organization according to config in org/
 # Must be run by a Github admin (generally in CI)
+echo "Writing generated org file to ${OUT_DIR}"
 
-/app/prow/cmd/peribolos/app.binary --fix-org --fix-org-members --fix-teams --fix-team-members \
-		--config-path "${WD}"/../org/istio.yaml --github-token-path /etc/github-token/oauth --confirm
+go run org/gen.go --output "${OUT_DIR}/istio.yaml"
+
+echo "Generated configuration: $(cat "${OUT_DIR}/istio.yaml")"
+
+PERIBOLOS="${PERIBOLOS:-peribolos}"
+# Fallback to legacy peribolos binary if not found
+if ! command -v "${PERIBOLOS}" &> /dev/null; then
+	PERIBOLOS=/app/prow/cmd/peribolos/app.binary
+fi
+
+"${PERIBOLOS}" --fix-org --fix-org-members --fix-teams --fix-team-members \
+	--config-path "${WD}"/../org/istio.yaml --github-token-path /etc/github-token/oauth --confirm
